Admin Pages

+++ front-end/src/pages/Account.tsx
@@ -3,8 +3,8 @@ import Header from "../components/layout/Header"
 import "./Account.css"
 import { UserContext } from "../App"
 import Button from "../components/Button"
-import { faAt, faPencil, faSave, faTrash, faUser } from "@fortawesome/free-solid-svg-icons"
-import { useNavigate } from "react-router-dom"
+import { faAt, faEnvelope, faPencil, faSave, faTrash, faUser } from "@fortawesome/free-solid-svg-icons"
+import { Link, useNavigate } from "react-router-dom"
 import type { UserUpdate } from "../server/ServerTypes"
 import server, { type ListingDetails } from "../server"
 import TextInput from "../components/forms/TextInput"
@@ -18,11 +18,6 @@ function Account() {
 	const navigate = useNavigate()
 
 	const [user, setUser] = useContext(UserContext)
-	if (!user) {
-		navigate("/")
-		return <></>
-	}
-	const currentUser = user
 
 	const [listings, setListings] = useState<ListingDetails[]>([])
 
@@ -33,27 +28,44 @@ function Account() {
 	const [emailErr, setEmailErr] = useState<string | undefined>(undefined)
 	const [updateFields, setUpdateFields] = useState<UserUpdate>({})
 
-	function updateListingView(): void {
-		const userId = currentUser.id
-		const fetchListings = async () => {
-			const [listings, _] = await server.searchListingsByUser(userId)
-			setListings(listings ?? [])
-		}
-		fetchListings()
-	}
+	const [pageLoadingMessage, setPageLoadingMessage] = useState(
+		<>Loading...</>
+	)
 
 	useEffect(() => {
-		const fetchListings = async () => {
-			const [listings, _] = await server.searchListingsByUser(currentUser.id)
-			setListings(listings ?? [])
+		// If after 2 seconds we haven't retrieved the user information, then
+		// we assume they're not logged in and prompt them to navigate to the
+		// login page.
+		setTimeout(() => {
+			setPageLoadingMessage(<>
+				Log in <Link to="/login">here</Link>.
+			</>)			
+		}, 2000)
+	}, [])
+
+	useEffect(() => {
+		if (user == null) {
+			return
 		}
-		fetchListings()
-	}, [currentUser.id])
+		server
+			.searchListingsByUser(user.id)
+			.then(([ listings, _ ]) => {
+				setListings(listings ?? [])
+			})
+	}, [user])
 
 	useEffect(() => {
 		setNameErr(undefined)
 		setEmailErr(undefined)
 	}, [editing])
+
+	// If the user information is pending, we render this.
+	if (user == null) {
+		return <>
+			<Header currentPage="/account" hideProfileMenu />
+			<p className="loadingMessage">{pageLoadingMessage}</p>
+		</>
+	}
 
 	return (<>
 		<Header currentPage="/account" hideProfileMenu />
@@ -86,13 +98,12 @@ function Account() {
 						/>
 						<TextInput
 							type="email"
-							icon={faAt}
+							icon={faEnvelope}
 							initialValue={user.email}
 							name="email"
 							error={emailErr}
 							onChange={email => {
 								setEmailErr(undefined)
-
 
 								if (email == user.email || email.length == 0) {
 									delete updateFields.email
@@ -187,6 +198,10 @@ function Account() {
 						text={"Delete Account"}
 						style={"important"}
 						onClick={() => {
+							if (!confirm("Are you sure you want to delete your account?")) {
+								return
+							}
+
 							setLoading(true)
 							server.deleteUser()
 							setLoading(false)
@@ -199,9 +214,24 @@ function Account() {
 			</section>
 			<section className="listings">
 				<h1>Your Listings</h1>
-				<div className="searchResultsContainer">
-					{listings.length > 0 ? listings.map(listing => <BookCardEditable book={listing.book} update={updateListingView} listingId={listing.id} available={listing.available}/>) : "No listings"}
-				</div>
+				{listings.length == 0
+					? <em className="noListingsMessage">No listings.</em>
+					: listings.map(listing => 
+						<BookCardEditable 
+							key={listing.id}
+							book={listing.book} 
+							onDelete={() => {
+								setListings(prev => {
+									return [...prev].filter(({ id }) => 
+										id  != listing.id
+									)
+								})
+							}}
+							listingId={listing.id} 
+							available={listing.available}
+						/>
+					)
+				}
 			</section>
 		</main>
 	</>)
